+++ packages/mui/src/ArrayFieldItemTemplate/ArrayFieldItemTemplate.tsx
@@ -36,11 +36,7 @@ export default function ArrayFieldItemTemplate<
     minWidth: 0,
   };
 
-  const muiProps = getMuiProps<T, S, F>({
-    uiSchema,
-    formContext: registry.formContext,
-    options: uiOptions,
-  });
+  const muiProps = getMuiProps<T, S, F>(uiOptions);
   const { slotProps: muiSlotProps, ...otherMuiProps } = muiProps;
 
   return (