+++ packages/mui/src/RangeWidget/RangeWidget.tsx
@@ -20,22 +20,8 @@ import { getMuiProps } from '../util';
 export default function RangeWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any>(
   props: WidgetProps<T, S, F>,
 ) {
-  const {
-    value,
-    readonly,
-    disabled,
-    onBlur,
-    onFocus,
-    options,
-    schema,
-    onChange,
-    required,
-    label,
-    hideLabel,
-    id,
-    registry,
-    uiSchema,
-  } = props;
+  const { value, readonly, disabled, onBlur, onFocus, options, schema, onChange, required, label, hideLabel, id } =
+    props;
   const sliderProps = { value, label, id, name: id, ...rangeSpec<S>(schema) };
 
   const _onChange = (_: any, value?: number | number[]) => {
@@ -44,12 +30,7 @@ export default function RangeWidget<T = any, S extends StrictRJSFSchema = RJSFSc
   const _onBlur = ({ target }: FocusEvent<HTMLInputElement>) => onBlur(id, target && target.value);
   const _onFocus = ({ target }: FocusEvent<HTMLInputElement>) => onFocus(id, target && target.value);
 
-  const muiProps = getMuiProps<T, S, F>({
-    uiSchema,
-    formContext: registry.formContext,
-    options,
-  });
-  const { slotProps: muiSlotProps, ...otherMuiProps } = muiProps;
+  const muiProps = getMuiProps<T, S, F, SliderProps>(options);
 
   return (
     <>
@@ -60,16 +41,13 @@ export default function RangeWidget<T = any, S extends StrictRJSFSchema = RJSFSc
         hideLabel,
       )}
       <Slider
-        {...(otherMuiProps as SliderProps)}
+        {...muiProps}
         disabled={disabled || readonly}
         onChange={_onChange}
         onBlur={_onBlur}
         onFocus={_onFocus}
         valueLabelDisplay='auto'
         {...sliderProps}
-        slotProps={{
-          ...muiSlotProps,
-        }}
         aria-describedby={ariaDescribedByIds(id)}
       />
     </>