+++ packages/mui/src/util.ts
@@ -1,42 +1,25 @@
-import {
-  FormContextType,
-  getUiOptions,
-  RJSFSchema,
-  StrictRJSFSchema,
-  UiSchema,
-  UIOptionsType,
-  GenericObjectType,
-} from '@rjsf/utils';
-
-export type MuiPropsType<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any> = {
-  formContext?: F;
-  uiSchema?: UiSchema<T, S, F>;
-  options?: UIOptionsType<T, S, F>;
-  defaultSchemaProps?: GenericObjectType;
-  defaultContextProps?: GenericObjectType;
-};
+import { FormContextType, RJSFSchema, StrictRJSFSchema, UIOptionsType, GenericObjectType } from '@rjsf/utils';
 
 /**
- * Extract props meant for MUI components from props that are
- * passed to Widgets, Templates and Fields.
- * @param {Object} params
- * @param {Object?} params.formContext
- * @param {Object?} params.uiSchema
- * @param {Object?} params.options
- * @param {Object?} params.defaultSchemaProps
- * @param {Object?} params.defaultContextProps
+ * Extract props meant for MUI components from the `options` field of the `uiSchema`.
+ * @param {UIOptionsType} options - The options from the uiSchema
+ * @param {string[]} [propsToFilter] - An optional list of props to filter out
  * @returns {any}
  */
-export function getMuiProps<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any>({
-  formContext = {} as F,
-  uiSchema = {},
-  options = {},
-  defaultSchemaProps = {},
-  defaultContextProps = {},
-}: MuiPropsType<T, S, F>) {
-  const FormContextProps = formContext.mui;
-  const SchemaProps = getUiOptions<T, S, F>(uiSchema).mui;
-  const OptionProps = options.mui;
-
-  return Object.assign({}, defaultSchemaProps, defaultContextProps, FormContextProps, SchemaProps, OptionProps);
+export function getMuiProps<
+  T = any,
+  S extends StrictRJSFSchema = RJSFSchema,
+  F extends FormContextType = any,
+  P extends GenericObjectType = GenericObjectType,
+>(options: UIOptionsType<T, S, F>, propsToFilter?: string[]): P {
+  const muiProps = (options?.mui as P) || ({} as P);
+  if (propsToFilter) {
+    return Object.keys(muiProps)
+      .filter((key) => propsToFilter.includes(key))
+      .reduce((obj, key) => {
+        obj[key as keyof P] = muiProps[key as keyof P];
+        return obj;
+      }, {} as P);
+  }
+  return muiProps;
 }